Charting. Move logic to hook files
