+++ src/components/RenderBundleResourceContext/index.tsx
@@ -1,27 +1,20 @@
 import { Bundle, Resource } from 'fhir/r4b';
-import { useParams } from 'react-router-dom';
 
-import { RenderRemoteData, useService, WithId } from '@beda.software/fhir-react';
+import { RenderRemoteData, WithId } from '@beda.software/fhir-react';
 
-import { getFHIRResources } from 'src/services';
 import { DetailPageProps } from 'src/uberComponents/ResourceDetailPage/types';
 import { RecordType } from 'src/uberComponents/ResourceListPage/types';
-import { compileAsFirst } from 'src/utils';
+
+import { useRenderBundleResourceContext } from './hooks';
 
 interface RenderBundleResourceContextProps<R extends Resource> extends DetailPageProps<R> {
     children: (context: RecordType<WithId<R>>) => JSX.Element;
 }
 
 export function RenderBundleResourceContext<R extends Resource>(props: RenderBundleResourceContextProps<R>) {
-    const { resourceType, getSearchParams, extractPrimaryResource, children } = props;
+    const { resourceType, extractPrimaryResource, children } = props;
 
-    const params = useParams();
-
-    const [response] = useService(() => getFHIRResources(resourceType, getSearchParams(params)));
-
-    const defaultExtractPrimaryResource = compileAsFirst<Bundle<WithId<R>>, WithId<R>>(
-        'Bundle.entry.resource.where(resourceType=%resourceType).first()',
-    );
+    const { response, defaultExtractPrimaryResource } = useRenderBundleResourceContext<R>(props);
 
     return (
         <RenderRemoteData remoteData={response}>
